patch selenium_test_case
