+++ test_settings.py
@@ -23,6 +23,8 @@ SELENIUM_WEBDRIVERS = {
     },
 }
 
+SELENIUM_WIDTHS = [1624]
+
 DATABASES = {
     "default": {
         "ENGINE": "django.db.backends.sqlite3",
